hello_fresh/requirements.txt
requests
brotli
pandas
scikit-learn
python-Levenshtein
gensim
transformers
torch
matplotlib
pillow
psutil